moved embedding out of search
