+++ tests/integration/dataset/search/test_embeddings.py
@@ -19,7 +19,7 @@ import pytest
 
 from kolena._experimental.search import upload_embeddings
 from kolena.dataset import upload_dataset
-from kolena.dataset.search.embeddings import _upload_dataset_embeddings
+from kolena.dataset.embeddings import _upload_dataset_embeddings
 from kolena.errors import InputValidationError
 from kolena.errors import NotFoundError
 from kolena.workflow import define_workflow
